refactor(ui): create l5r/ui/ package and extract widget helpers
First step of the Phase 4 main.py split. Adds l5r/ui/ (and
l5r/ui/tabs/) as empty packages, and moves the five module-level
widget factories out of main.py into l5r/ui/helpers.py:

* new_small_le
* new_horiz_line
* new_vert_line
* new_item_groupbox
* new_small_plus_bt

main.py re-imports them at module level so the existing call sites
inside L5RMain methods keep working unchanged. No behavior change.

main.py: 2551 -> 2503 lines.
